changed timestamp to long to avoid overflow

+++ Services/RawDataToClientData/RawDataToClientData/Models/DroneSensors.cs
@@ -93,7 +93,7 @@ namespace RawDataToClientData
         }
 
 
-        private static bool IsActive(int usvTimestamp)
+        private static bool IsActive(long usvTimestamp)
         {
             var oneHourMilliseconds = 3600000;
             var timestamp = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() - oneHourMilliseconds;
